separate start stream stuff from index page
