+++ frontend/src/index.css
@@ -73,17 +73,19 @@ button:focus-visible {
 
 */
 
+/* Index page */
+
 .streamList {
   position: relative;
-  margin-left: auto;
-  margin-right: auto;
+  margin-left: 2rem;
+  margin-right: 2rem;
   background-color: rgba(2, 2, 2, 0.397);
   overflow-y: auto;
   overflow-x: hidden;
   display: flex;
   flex-direction: column;
   gap: 0.5rem;
-  padding: 1rem;
+  padding: 2rem;
   border-radius: 2rem;
   user-select: none;
 }
@@ -97,11 +99,31 @@ button:focus-visible {
   background-color: rgba(255, 255, 255, 0.05);
   border-radius: 2rem;
   transition: 200ms;
+  padding: 1rem;
 }
 
 .streamListItem:hover {
   background-color: rgba(255, 255, 255, 0.1);
 }
+
+.header {
+  padding: 1rem;
+  color: white;
+}
+
+.main {
+  position: fixed;
+  top: 0px;
+  left: 0px;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+
+
+/* Viewer/streamer */
 
 .statusIndicator {
   position: fixed;
@@ -135,19 +157,4 @@ button:focus-visible {
   height: 16px;
   border-radius: 16px;
   opacity: 0.3;
-}
-
-.header {
-  padding: 1rem;
-  color: white;
-}
-
-.main {
-  position: fixed;
-  top: 0px;
-  left: 0px;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
 }
